+++ include/roq/application.h
@@ -2,6 +2,9 @@
 
 #pragma once
 
+#include <string>
+#include <string_view>
+
 #include "roq/logging.h"
 
 namespace roq {
@@ -13,7 +16,11 @@ class ROQ_LOGGING_PUBLIC Application {
       int argc,
       char **argv,
       const std::string_view& description,
-      const std::string_view& version);
+      const std::string_view& version,
+      const std::string_view& build_type = std::string_view(),
+      const std::string_view& git_hash = std::string_view(),
+      const std::string_view& compile_date = __DATE__,
+      const std::string_view& compile_time = __TIME__);
 
   Application(const Application&) = delete;
   Application(Application&&) = default;
@@ -29,6 +36,10 @@ class ROQ_LOGGING_PUBLIC Application {
  private:
   char **_argv;
   int _argc;
+  std::string _build_type;
+  std::string _git_hash;
+  std::string _compile_date;
+  std::string _compile_time;
 };
 
 }  // namespace roq