+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Added
+
+* Application will now log process information on startup
+
 ## 0.4.2 &ndash; 2020-07-27
 
 ### Removed